Pick up exec from path, now an auto-test.

+++ test/search_path_test_2.sh
@@ -1,10 +1,8 @@
 #!/bin/bash
 
-EXEC_DIR=${1:-.}
-
 export TMP_DIR=`mktemp -d /tmp/${0##*/}.XXXXXXXXXX`
 
-trap "[[ -d \"$TMP_DIR\" ]] && rm -rf \"$TMP_DIR\"" EXIT
+[[ -z "$DEBUG" ]] && trap "[[ -d \"$TMP_DIR\" ]] && rm -rf \"$TMP_DIR\"" EXIT
 
 T1="libthis_is_one.so libanother_one.so"
 T2="libthis_is_two.so libanother_two.so"
@@ -18,9 +16,9 @@ done
 
 echo `pwd`
 
-${EXEC_DIR}/search_path_test_2 "lib.*one.so"   TMP_DIR; test $? -eq 2 || exit 1
-${EXEC_DIR}/search_path_test_2 "lib.*two.so"   TMP_DIR; test $? -eq 2 || exit 2
-${EXEC_DIR}/search_path_test_2 "lib.*three.so" TMP_DIR; test $? -eq 1 || exit 3
-${EXEC_DIR}/search_path_test_2 "lib.*four.so"  TMP_DIR; test $? -eq 0 || exit 4
+search_path_test_2 "lib.*one.so"   TMP_DIR; test $? -eq 2 || exit 1
+search_path_test_2 "lib.*two.so"   TMP_DIR; test $? -eq 2 || exit 2
+search_path_test_2 "lib.*three.so" TMP_DIR; test $? -eq 1 || exit 3
+search_path_test_2 "lib.*four.so"  TMP_DIR; test $? -eq 0 || exit 4
 
 exit 0
